requirements.txt
openai==0.28.1
pandas>=1.3.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0
schedule>=1.1.0
watchdog>=2.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11